08_11656_접미사 배열

+++ Solution/Solution/Baekjoon_Online_Judge/00. code.plus/00_알고리즘 기초/01_자료구조 1/04_참고/07_10824_네 수.cpp
@@ -17,7 +17,7 @@ int main()
 	for (string& strNum : arrNumBuff)
 		cin >> strNum;
 
-	cout << stoi(arrNumBuff[NUM_A] + arrNumBuff[NUM_B]) + stoi(arrNumBuff[NUM_C] + arrNumBuff[NUM_D]);
+	cout << stoll(arrNumBuff[NUM_A] + arrNumBuff[NUM_B]) + stoll(arrNumBuff[NUM_C] + arrNumBuff[NUM_D]);
 
 	return 0;
 }
